+++ simplenote/__init__.py
@@ -3,5 +3,5 @@
 from simplenote import Simplenote
 
 __author__ = "Daniel Schauenberg"
-__version__ = "0.1.3"
+__version__ = "0.2.0"
 __license__ = "MIT"